requirements.txt
flask==3.0.2
gunicorn==21.2.0
transformers==4.43.4
torch==2.3.1
faiss-cpu==1.8.0
numpy==1.26.4
